tereverde/requirements.txt
Django
djongo
djangorestframework
pytest-django
pytz
Pillow
django-multiselectfield